Crate ln_gateway

Modules

client

db 🔒

envs

gateway_lnrpc

gateway_module_v2

lightning

rpc

state_machine

types 🔒

Map gateway_lnrpc protobuf types to rust types

Structs

ClientsJoinLock 🔒

Gateway

GatewayOpts 🔒

Command line parameters for starting the gateway. mode, data_dir, listen, and api_addr are all required.

GatewayParameters

GatewayParameters is a helper struct that can be derived from GatewayOpts that holds the CLI or environment variables that are specified by the user.

LightningContext

Represents an active connection to the lightning node.

PrettyInterceptHtlcRequest 🔒

Utility struct for formatting an intercepted HTLC. Useful for debugging.

Enums

GatewayError

Errors that can occur while processing incoming HTLC’s, making outgoing payments, registering with connected federations, or responding to webserver requests.

GatewayState