ln_gateway
Module envs

Constants

FM_DEBUG_GATEWAY_ENV

Environment variable that instructs the gateway to run in “debug mode”, which allows errors to return to clients without redacting private information.

FM_GATEWAY_API_ADDR_ENV

Environment variable that specifies the URL that clients can use to make requests to the gateway.

FM_GATEWAY_BCRYPT_PASSWORD_HASH_ENV

Environment variable that specifies the bcrypt password hash.

FM_GATEWAY_DATA_DIR_ENV

Environment variable that specifies the directory of the gateway’s database.

FM_GATEWAY_LIGHTNING_MODULE_MODE_ENV

Environment variable that specifies the “module mode” the gateway should run in. Options are “LNv1”, “LNv2”, or “All”. It is not recommended to run “All” in production so that clients are not able to use the same gateway to create LNv1 and LNv2 invoices.

FM_GATEWAY_LISTEN_ADDR_ENV

Environment variable that specifies the address the gateway’s HTTP server should listen on.

FM_GATEWAY_MNEMONIC_ENV

Environment variable that specifies the mnemonic that the gateway should use for ecash and the LDK Node should use for onchain funds. If not set, a mnemonic will be generated. This environment variable can be used for recovering from an existing mnemonic.

FM_GATEWAY_NETWORK_ENV

Environment variable that specifies that Bitcoin network that the gateway should use. Must match the network of the Lightning node.

FM_GATEWAY_SKIP_WAIT_FOR_SYNC_ENV

FM_LDK_ESPLORA_SERVER_URL

Environment variable that specifies the URL of an Esplora server. Necessary for LDK configuration.

FM_LDK_NETWORK

Environment variable that specifies the Bitcoin network that the LDK Node should use. Must match FM_GATEWAY_NETWORK. Necessary for LDK configuration.

FM_LND_MACAROON_ENV

Environment variable that specifies the location of LND’s macaroon. Necessary for LND configuration.

FM_LND_RPC_ADDR_ENV

Environment variable that specifies the URL to connect to LND. Necessary for LND configuration.

FM_LND_TLS_CERT_ENV

Environment variable that specifies the location of LND’s TLS certificate. Necessary for LND configuration.

FM_NUMBER_OF_ROUTE_HINTS_ENV

Environment variable that instructs the gateway how many route hints to include in LNv1 invoices.

FM_PORT_LDK

Environment variable the specifies the port that the LDK Node should use. Necessary for LDK configuration.