ln_gateway::gateway_lnrpc::gateway_lightning_client
Struct GatewayLightningClient

pub struct GatewayLightningClient<T> {
    inner: Grpc<T>,
}

GatewayLightning is a service that provides limited access and functionality from a lightning node to Fedimint gateways

Fields
inner: Grpc<T>
Implementations

impl GatewayLightningClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> GatewayLightningClient<T>
where T: GrpcService<BoxBody>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> GatewayLightningClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<BoxBody>, Response = Response<<T as GrpcService<BoxBody>>::ResponseBody>>, <T as Service<Request<BoxBody>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn get_node_info( &mut self, request: impl IntoRequest<EmptyRequest>, ) -> Result<Response<GetNodeInfoResponse>, Status>

GetNodeInfo returns the public key and alias of the associated lightning node

pub async fn get_route_hints( &mut self, request: impl IntoRequest<GetRouteHintsRequest>, ) -> Result<Response<GetRouteHintsResponse>, Status>

GetRouteHints returns the route hints to the associated lightning node

pub async fn pay_invoice( &mut self, request: impl IntoRequest<PayInvoiceRequest>, ) -> Result<Response<PayInvoiceResponse>, Status>

PayInvoice attempts to pay an invoice using the associated lightning node

pub async fn pay_pruned_invoice( &mut self, request: impl IntoRequest<PayPrunedInvoiceRequest>, ) -> Result<Response<PayInvoiceResponse>, Status>

PayPrunedInvoice attempts to pay a pruned invoice, which is an invoice without a description, using the associated lightning node.

pub async fn route_htlcs( &mut self, request: impl IntoRequest<EmptyRequest>, ) -> Result<Response<Streaming<InterceptHtlcRequest>>, Status>

RouteHtlcs opens a bi-directional stream for the client to receive intercepted HTLCs. InterceptHtlcRequest is sent from the server to alert the client that an HTLC has been intercepted and needs to be processed. The client is expected to response with InterceptHtlcResponse after the HTLC has been processed with the appropriate action (Settle, Forward, Cancel).

pub async fn complete_htlc( &mut self, request: impl IntoRequest<InterceptHtlcResponse>, ) -> Result<Response<EmptyResponse>, Status>

pub async fn create_invoice( &mut self, request: impl IntoRequest<CreateInvoiceRequest>, ) -> Result<Response<CreateInvoiceResponse>, Status>

pub async fn get_ln_onchain_address( &mut self, request: impl IntoRequest<EmptyRequest>, ) -> Result<Response<GetLnOnchainAddressResponse>, Status>

Get a Bitcoin address that belongs to the underlying lightning node’s wallet.

pub async fn withdraw_onchain( &mut self, request: impl IntoRequest<WithdrawOnchainRequest>, ) -> Result<Response<WithdrawOnchainResponse>, Status>

pub async fn open_channel( &mut self, request: impl IntoRequest<OpenChannelRequest>, ) -> Result<Response<OpenChannelResponse>, Status>

Open a channel on the underlying lightning node.

pub async fn close_channels_with_peer( &mut self, request: impl IntoRequest<CloseChannelsWithPeerRequest>, ) -> Result<Response<CloseChannelsWithPeerResponse>, Status>

Close all channels with a given peer on the underlying lightning node.

pub async fn list_active_channels( &mut self, request: impl IntoRequest<EmptyRequest>, ) -> Result<Response<ListActiveChannelsResponse>, Status>

List all channels that are active and able to send and receive funds.

pub async fn get_balances( &mut self, request: impl IntoRequest<EmptyRequest>, ) -> Result<Response<GetBalancesResponse>, Status>