Trait ln_gateway::gatewaylnrpc::gateway_lightning_server::GatewayLightning

pub trait GatewayLightning: Send + Sync + 'static {
    type SubscribeInterceptHtlcsStream: Stream<Item = Result<SubscribeInterceptHtlcsResponse, Status>> + Send + 'static;

    fn get_pub_key<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetPubKeyRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPubKeyResponse>, Status>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn pay_invoice<'life0, 'async_trait>(
        &'life0 self,
        request: Request<PayInvoiceRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PayInvoiceResponse>, Status>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_intercept_htlcs<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscribeInterceptHtlcsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeInterceptHtlcsStream>, Status>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn complete_htlc<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CompleteHtlcsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<CompleteHtlcsResponse>, Status>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with GatewayLightningServer.

Required Associated Types

type SubscribeInterceptHtlcsStream: Stream<Item = Result<SubscribeInterceptHtlcsResponse, Status>> + Send + 'static

Server streaming response type for the SubscribeInterceptHtlcs method.

Required Methods

fn get_pub_key<'life0, 'async_trait>(
&'life0 self,
request: Request<GetPubKeyRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetPubKeyResponse>, Status>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

GetPubKey returns the public key of the associated lightning node

fn pay_invoice<'life0, 'async_trait>(
&'life0 self,
request: Request<PayInvoiceRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<PayInvoiceResponse>, Status>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

PayInvoice attempts to pay an invoice using the associated lightning node

fn subscribe_intercept_htlcs<'life0, 'async_trait>(
&'life0 self,
request: Request<SubscribeInterceptHtlcsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeInterceptHtlcsStream>, Status>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

SubscribeInterceptHtlcs opens a stream for a client to receive specific HTLCs that have a specific short channel id. For every HTLC intercepted and processed, the client should use CompleteHtlcs RPC to stream back a Success or Failure response.

Recommendation: GatewayLightning implementations should respond with a channel stream over which intercepted HTLCs are continually sent to the client.

fn complete_htlc<'life0, 'async_trait>(
&'life0 self,
request: Request<CompleteHtlcsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<CompleteHtlcsResponse>, Status>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

CompleteHtlc allows a client to send a Success or Failure response for a HTLC that was intercepted and processed.