Module ln_gateway::lightning

Modules

cln

ldk

lnd

Structs

ChannelInfo

GatewayLightningBuilder

LightningContext

Represents an active connection to the lightning node.

Enums

LightningMode

LightningRpcError

Constants

MAX_LIGHTNING_RETRIES

Traits

ILnRpcClient

A trait that the gateway uses to interact with a lightning node. This allows the gateway to be agnostic to the specific lightning node implementation being used.

LightningBuilder

Type Aliases

HtlcResult

RouteHtlcStream