ln_gateway::lightning::ldk
Struct GatewayLdkClient

pub struct GatewayLdkClient {
    node: Arc<Node>,
    esplora_client: AsyncClient,
    event_handler_task_handle: JoinHandle<()>,
    htlc_stream_receiver_or: Option<Receiver<Result<InterceptHtlcRequest, Status>>>,
}

Fields
node: Arc<Node>

The underlying lightning node.

esplora_client: AsyncClient

The client for querying data about the blockchain.

event_handler_task_handle: JoinHandle<()>

A handle to the task that processes incoming events from the lightning node. Responsible for sending incoming HTLCs to the caller of route_htlcs. TODO: This should be a shutdown sender instead, and we can discard the handle.

htlc_stream_receiver_or: Option<Receiver<Result<InterceptHtlcRequest, Status>>>

The HTLC stream, until it is taken by calling ILnRpcClient::route_htlcs.

Implementations

impl GatewayLdkClient

pub fn new( data_dir: &Path, esplora_server_url: &str, network: Network, lightning_port: u16, mnemonic: Mnemonic, ) -> Result<Self>

Creates a new GatewayLdkClient instance and starts the underlying lightning node. All resources, including the lightning node, will be cleaned up when the returned GatewayLdkClient instance is dropped. There’s no need to manually stop the node.

async fn handle_next_event( node: &Node, htlc_stream_sender: &Sender<Result<InterceptHtlcRequest, Status>>, )

async fn outpoint_to_scid(&self, funding_txo: OutPoint) -> Result<u64>

Converts a transaction outpoint to a short channel ID by querying the blockchain.