Trait ln_gateway::lnrpc_client::ILnRpcClient

pub trait ILnRpcClient: Debug + Send + Sync {
    fn pubkey<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<GetPubKeyResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn routehints<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<GetRouteHintsResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn pay<'life0, 'async_trait>(
        &'life0 self,
        invoice: PayInvoiceRequest
    ) -> Pin<Box<dyn Future<Output = Result<PayInvoiceResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn subscribe_htlcs<'a, 'life0, 'async_trait>(
        &'life0 self,
        subscription: SubscribeInterceptHtlcsRequest
    ) -> Pin<Box<dyn Future<Output = Result<HtlcStream<'a>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'a: 'async_trait,
        'life0: 'async_trait;
    fn complete_htlc<'life0, 'async_trait>(
        &'life0 self,
        outcome: CompleteHtlcsRequest
    ) -> Pin<Box<dyn Future<Output = Result<CompleteHtlcsResponse>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
}

Required Methods

fn pubkey<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<GetPubKeyResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Get the public key of the lightning node

fn routehints<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<GetRouteHintsResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Get route hints to the lightning node

fn pay<'life0, 'async_trait>(
&'life0 self,
invoice: PayInvoiceRequest
) -> Pin<Box<dyn Future<Output = Result<PayInvoiceResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Attempt to pay an invoice using the lightning node

fn subscribe_htlcs<'a, 'life0, 'async_trait>(
&'life0 self,
subscription: SubscribeInterceptHtlcsRequest
) -> Pin<Box<dyn Future<Output = Result<HtlcStream<'a>>> + Send + 'async_trait>>where
Self: 'async_trait,
'a: 'async_trait,
'life0: 'async_trait,

Subscribe to intercept htlcs that belong to a specific mint identified by short_channel_id

fn complete_htlc<'life0, 'async_trait>(
&'life0 self,
outcome: CompleteHtlcsRequest
) -> Pin<Box<dyn Future<Output = Result<CompleteHtlcsResponse>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Request completion of an intercepted htlc after processing and determining an outcome