Struct ln_gateway::rpc::rpc_client::Error

pub struct Error {
    inner: Box<Inner, Global>,
}

The Errors that may occur when processing a Request.

Note: Errors may include the full URL used to make the Request. If the URL contains sensitive information (e.g. an API key as a query parameter), be sure to remove it (without_url)

Fields
inner: Box<Inner, Global>
Implementations

impl Error

pub fn url(&self) -> Option<&Url>

Returns a possible URL related to this error.

Examples

// displays last stop of a redirect loop
let response = reqwest::get("http://site.with.redirect.loop").await;
if let Err(e) = response {
    if e.is_redirect() {
        if let Some(final_stop) = e.url() {
            println!("redirect loop at {}", final_stop);
        }
    }
}

pub fn url_mut(&mut self) -> Option<&mut Url>

Returns a mutable reference to the URL related to this error

This is useful if you need to remove sensitive information from the URL (e.g. an API key in the query), but do not want to remove the URL entirely.

pub fn with_url(self, url: Url) -> Error

Add a url related to this error (overwriting any existing)

pub fn without_url(self) -> Error

Strip the related url from this error (if, for example, it contains sensitive information)

pub fn is_builder(&self) -> bool

Returns true if the error is from a type Builder.

pub fn is_redirect(&self) -> bool

Returns true if the error is from a RedirectPolicy.

pub fn is_status(&self) -> bool

Returns true if the error is from Response::error_for_status.

pub fn is_timeout(&self) -> bool

Returns true if the error is related to a timeout.

pub fn is_request(&self) -> bool

Returns true if the error is related to the request

pub fn is_connect(&self) -> bool

Returns true if the error is related to connect

pub fn is_body(&self) -> bool

Returns true if the error is related to the request or response body

pub fn is_decode(&self) -> bool

Returns true if the error is related to decoding the responseโs body

pub fn status(&self) -> Option<StatusCode>

Returns the status code, if the error was generated from a response.