ln_gateway::state_machine
Module complete

Structs

CompleteHtlcState

GatewayCompleteCommon

GatewayCompleteStateMachine

WaitForPreimageState

Enums

CompleteHtlcError 🔒

GatewayCompleteStates

State machine that completes the incoming payment by contacting the lightning node when the incoming contract has been funded and the preimage is available.

HtlcOutcome 🔒