ln_gateway::state_machine
Module events

Structs

CompleteLightningPaymentSucceeded

LNv1 event that is emitted when a preimage was successfully revealed to the Lightning Network.

IncomingPaymentFailed

LNv1 event that is emitted when an incoming payment attempt has failed.

IncomingPaymentStarted

LNv1 event that is emitted when an incoming payment attempt has started.

IncomingPaymentSucceeded

LNv1 event that is emitted when an incoming payment attempt was successful.

OutgoingPaymentFailed

LNv1 event that is emitted when an outgoing payment attempt has failed.

OutgoingPaymentStarted

LNv1 event that is emitted when an outgoing payment attempt is initiated.

OutgoingPaymentSucceeded

LNv1 event that is emitted when an outgoing payment attempt has succeeded.