Struct ln_gateway::state_machine::GatewayClientModule

pub struct GatewayClientModule {
    cfg: LightningClientConfig,
    pub notifier: ModuleNotifier<GatewayClientStateMachines>,
    pub redeem_key: KeyPair,
    timelock_delta: u64,
    federation_index: u64,
    module_api: DynModuleApi,
    client_ctx: ClientContext<Self>,
    gateway: Arc<Gateway>,
}

Client side Lightning module for the gateway.

For the client side Lightning module for normal clients, see fedimint_ln_client::LightningClientModule

Fields
cfg: LightningClientConfignotifier: ModuleNotifier<GatewayClientStateMachines>redeem_key: KeyPairtimelock_delta: u64federation_index: u64module_api: DynModuleApiclient_ctx: ClientContext<Self>gateway: Arc<Gateway>
Implementations

impl GatewayClientModule

fn to_gateway_registration_info( &self, route_hints: Vec<RouteHint>, ttl: Duration, fees: RoutingFees, lightning_context: LightningContext, ) -> LightningGatewayAnnouncement

async fn create_funding_incoming_contract_output_from_htlc( &self, htlc: Htlc, ) -> Result<(OperationId, Amount, ClientOutput<LightningOutputV0, GatewayClientStateMachines>), IncomingSmError>

async fn create_funding_incoming_contract_output_from_swap( &self, swap: SwapParameters, ) -> Result<(OperationId, ClientOutput<LightningOutputV0, GatewayClientStateMachines>), IncomingSmError>

pub async fn register_with_federation( &self, route_hints: Vec<RouteHint>, time_to_live: Duration, fees: RoutingFees, lightning_context: LightningContext, ) -> Result<()>

Register gateway with federation

pub async fn remove_from_federation(&self, gateway_keypair: KeyPair)

Attempts to remove a gateway’s registration from the federation. Since removing gateway registrations is best effort, this does not return an error and simply emits a warning when the registration cannot be removed.

async fn remove_from_federation_inner( &self, gateway_keypair: KeyPair, ) -> Result<()>

Retrieves the signing challenge from each federation peer. Since each peer maintains their own list of registered gateways, the gateway needs to provide a signature that is signed by the private key of the gateway id to remove the registration.

pub async fn gateway_handle_intercepted_htlc( &self, htlc: Htlc, ) -> Result<OperationId>

Attempt fulfill HTLC by buying preimage from the federation

async fn gateway_handle_direct_swap( &self, swap_params: SwapParameters, ) -> Result<OperationId>

Attempt buying preimage from this federation in order to fulfill a pay request in another federation served by this gateway. In direct swap scenario, the gateway DOES NOT send payment over the lightning network

pub async fn gateway_subscribe_ln_receive( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<GatewayExtReceiveStates>>

Subscribe to updates when the gateway is handling an intercepted HTLC, or direct swap between federations

pub async fn await_completion(&self, operation_id: OperationId)

For the given OperationId, this function will wait until the Complete state machine has finished or failed.

pub async fn gateway_pay_bolt11_invoice( &self, pay_invoice_payload: PayInvoicePayload, ) -> Result<OperationId>

Pay lightning invoice on behalf of federation user

pub async fn gateway_subscribe_ln_pay( &self, operation_id: OperationId, ) -> Result<UpdateStreamOrOutcome<GatewayExtPayStates>>