Trait ln_gateway::state_machine::GatewayClientExt

pub trait GatewayClientExt {
    // Required methods
    fn gateway_pay_bolt11_invoice<'life0, 'async_trait>(
        &'life0 self,
        pay_invoice_payload: PayInvoicePayload
    ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gateway_subscribe_ln_pay<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<GatewayExtPayStates>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_with_federation<'life0, 'async_trait>(
        &'life0 self,
        gateway_api: SafeUrl,
        route_hints: Vec<RouteHint>,
        time_to_live: Duration,
        gateway_id: PublicKey
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gateway_handle_intercepted_htlc<'life0, 'async_trait>(
        &'life0 self,
        htlc: Htlc
    ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gateway_handle_direct_swap<'life0, 'async_trait>(
        &'life0 self,
        swap_params: SwapParameters
    ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn gateway_subscribe_ln_receive<'life0, 'async_trait>(
        &'life0 self,
        operation_id: OperationId
    ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<GatewayExtReceiveStates>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn gateway_pay_bolt11_invoice<'life0, 'async_trait>( &'life0 self, pay_invoice_payload: PayInvoicePayload ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Pay lightning invoice on behalf of federation user

fn gateway_subscribe_ln_pay<'life0, 'async_trait>( &'life0 self, operation_id: OperationId ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<GatewayExtPayStates>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Subscribe to updates when the gateway is paying an invoice

fn register_with_federation<'life0, 'async_trait>( &'life0 self, gateway_api: SafeUrl, route_hints: Vec<RouteHint>, time_to_live: Duration, gateway_id: PublicKey ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Register gateway with federation

fn gateway_handle_intercepted_htlc<'life0, 'async_trait>( &'life0 self, htlc: Htlc ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Attempt fulfill HTLC by buying preimage from the federation

fn gateway_handle_direct_swap<'life0, 'async_trait>( &'life0 self, swap_params: SwapParameters ) -> Pin<Box<dyn Future<Output = Result<OperationId>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Attempt buying preimage from this federation in order to fulfill a pay request in another federation served by this gateway. In direct swap scenario, the gateway DOES NOT send payment over the lightning network

fn gateway_subscribe_ln_receive<'life0, 'async_trait>( &'life0 self, operation_id: OperationId ) -> Pin<Box<dyn Future<Output = Result<UpdateStreamOrOutcome<GatewayExtReceiveStates>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Subscribe to updates when the gateway is handling an intercepted HTLC, or direct swap between federations