Struct ln_gateway::Gateway

pub struct Gateway {
    pub(crate) federation_manager: Arc<RwLock<FederationManager>>,
    pub(crate) lightning_builder: Arc<dyn LightningBuilder + Send + Sync>,
    pub(crate) gateway_config: Arc<RwLock<Option<GatewayConfiguration>>>,
    pub(crate) state: Arc<RwLock<GatewayState>>,
    pub(crate) client_builder: GatewayClientBuilder,
    pub(crate) gateway_db: Database,
    pub(crate) client_joining_lock: Arc<Mutex<ClientsJoinLock>>,
    pub(crate) gateway_id: PublicKey,
    pub(crate) versioned_api: SafeUrl,
    pub(crate) listen: SocketAddr,
}

Fields
federation_manager: Arc<RwLock<FederationManager>>

The gateway’s federation manager.

lightning_builder: Arc<dyn LightningBuilder + Send + Sync>

Builder struct that allows the gateway to build a ILnRpcClient, which represents a connection to a lightning node.

gateway_config: Arc<RwLock<Option<GatewayConfiguration>>>

The gateway’s current configuration

state: Arc<RwLock<GatewayState>>

The current state of the Gateway.

client_builder: GatewayClientBuilder

Builder struct that allows the gateway to build a Fedimint client, which handles the communication with a federation.

gateway_db: Database

Database for Gateway metadata.

client_joining_lock: Arc<Mutex<ClientsJoinLock>>

Joining or leaving Federation is protected by this lock to prevent trying to use same database at the same time from multiple threads. Could be more granular (per id), but shouldn’t matter in practice.

gateway_id: PublicKey

A public key representing the identity of the gateway. Private key is not used.

versioned_api: SafeUrl

The Gateway’s API URL.

listen: SocketAddr

The socket the gateway listens on.

Implementations

impl Gateway

pub async fn new_with_custom_registry( lightning_builder: Arc<dyn LightningBuilder + Send + Sync>, client_builder: GatewayClientBuilder, listen: SocketAddr, api_addr: SafeUrl, cli_password: Option<String>, network: Option<Network>, fees: RoutingFees, num_route_hints: u32, gateway_db: Database, gateway_state: GatewayState, ) -> Result<Gateway>

Creates a new gateway but with a custom module registry provided inside client_builder. Currently only used for testing.

pub async fn new_with_default_modules() -> Result<Gateway>

Default function for creating a gateway with the Mint, Wallet, and Gateway modules.

pub(crate) async fn new( lightning_builder: Arc<dyn LightningBuilder + Send + Sync>, gateway_parameters: GatewayParameters, gateway_db: Database, client_builder: GatewayClientBuilder, gateway_state: GatewayState, ) -> Result<Gateway>

Helper function for creating a gateway from either new_with_default_modules or new_with_custom_registry.

pub(crate) async fn load_or_create_gateway_id( gateway_db: &Database, ) -> PublicKey

Returns a PublicKey that uniquely identifies the Gateway.

pub fn gateway_id(&self) -> PublicKey

pub fn versioned_api(&self) -> &SafeUrl

pub async fn clone_gateway_config(&self) -> Option<GatewayConfiguration>

pub async fn get_state(&self) -> GatewayState

pub async fn dump_database( dbtx: &mut DatabaseTransaction<'_>, prefix_names: Vec<String>, ) -> BTreeMap<String, Box<dyn Serialize + Send>>

Reads and serializes structures from the Gateway’s database for the purpose for serializing to JSON for inspection.

pub async fn run(self, tg: &TaskGroup) -> Result<TaskShutdownToken>

Main entrypoint into the gateway that starts the client registration timer, loads the federation clients from the persisted config, begins listening for intercepted HTLCs, and starts the webserver to service requests.

pub(crate) fn start_gateway(&self, task_group: &TaskGroup)

Begins the task for listening for intercepted HTLCs from the Lightning node.

pub(crate) async fn handle_disconnect(&self, htlc_task_group: TaskGroup)

Utility function for waiting for the task that is listening for intercepted HTLCs to shutdown.

pub async fn handle_htlc_stream( &self, stream: RouteHtlcStream<'_>, handle: TaskHandle, )

Blocks waiting for intercepted HTLCs to be sent over the stream. Spawns a state machine to either forward, cancel, or complete the HTLC depending on if the gateway is able to acquire the preimage from the federation.

pub(crate) async fn set_gateway_state(&self, state: GatewayState)

Helper function for atomically changing the Gateway’s internal state.

pub async fn handle_get_info(&self) -> Result<GatewayInfo>

Returns information about the Gateway back to the client when requested via the webserver.

pub async fn handle_get_federation_config( &self, federation_id_or: Option<FederationId>, ) -> Result<GatewayFedConfig>

If the Gateway is connected to the Lightning node, returns the ClientConfig for each federation that the Gateway is connected to.

pub async fn handle_balance_msg( &self, payload: BalancePayload, ) -> Result<Amount>

Returns the balance of the requested federation that the Gateway is connected to.

pub async fn handle_address_msg( &self, payload: DepositAddressPayload, ) -> Result<Address>

Returns a Bitcoin deposit on-chain address for pegging in Bitcoin for a specific connected federation.

pub async fn handle_withdraw_msg( &self, payload: WithdrawPayload, ) -> Result<Txid>

Returns a Bitcoin TXID from a peg-out transaction for a specific connected federation.

pub(crate) async fn handle_pay_invoice_msg( &self, payload: PayInvoicePayload, ) -> Result<Preimage>

Requests the gateway to pay an outgoing LN invoice on behalf of a Fedimint client. Returns the payment hash’s preimage on success.

pub async fn handle_connect_federation( &self, payload: ConnectFedPayload, ) -> Result<FederationInfo>

Handles a connection request to join a new federation. The gateway will download the federation’s client configuration, construct a new client, registers, the gateway with the federation, and persists the necessary config to reconstruct the client when restarting the gateway.

pub async fn handle_leave_federation( &self, payload: LeaveFedPayload, ) -> Result<FederationInfo>

Handle a request to have the Gateway leave a federation. The Gateway will request the federation to remove the registration record and the gateway will remove the configuration needed to construct the federation client.

pub fn handle_backup_msg(&self, _: BackupPayload) -> Result<()>

Handles a request for the gateway to backup a connected federation’s ecash. Not currently supported.

pub fn handle_restore_msg(&self, _: RestorePayload) -> Result<()>

Handles a request for the gateway to restore a connected federation’s ecash. Not currently supported.

pub async fn handle_set_configuration_msg( &self, __arg1: SetConfigurationPayload, ) -> Result<()>

Handle a request to change a connected federation’s configuration or gateway metadata. If num_route_hints is changed, the Gateway will re-register with all connected federations. If per_federation_routing_fees is changed, the Gateway will only re-register with the specified federation.

pub async fn handle_get_funding_address_msg(&self) -> Result<Address>

pub async fn handle_open_channel_msg( &self, __arg1: OpenChannelPayload, ) -> Result<()>

Instructs the Gateway’s Lightning node to open a channel to a peer specified by pubkey.

pub async fn handle_close_channels_with_peer_msg( &self, __arg1: CloseChannelsWithPeerPayload, ) -> Result<CloseChannelsWithPeerResponse>

Instructs the Gateway’s Lightning node to close all channels with a peer specified by pubkey.

pub async fn handle_list_active_channels_msg(&self) -> Result<Vec<ChannelInfo>>

Returns a list of Lightning network channels from the Gateway’s Lightning node.

pub(crate) async fn register_federations( &self, gateway_config: &GatewayConfiguration, federations: &[(FederationId, FederationConfig)], ) -> Result<()>

Registers the gateway with each specified federation.

pub(crate) async fn get_gateway_configuration( gateway_db: Database, gateway_parameters: &GatewayParameters, ) -> Option<GatewayConfiguration>

This function will return a GatewayConfiguration one of two ways. To avoid conflicting configs, the below order is the order in which the gateway will respect configurations:

GatewayConfiguration is read from the database.

All cli or environment variables are set such that we can create a GatewayConfiguration

pub async fn select_client( &self, federation_id: FederationId, ) -> Result<Spanned<ClientHandleArc>>

Retrieves a ClientHandleArc from the Gateway’s in memory structures that keep track of available clients, given a federation_id.

pub(crate) async fn load_clients(&self)

Reads the connected federation client configs from the Gateway’s database and reconstructs the clients necessary for interacting with connection federations.

pub(crate) fn register_clients_timer(&self, task_group: &TaskGroup)

Legacy mechanism for registering the Gateway with connected federations. This will spawn a task that will re-register the Gateway with connected federations every 8.5 mins. Only registers the Gateway if it has successfully connected to the Lightning node, so that it can include route hints in the registration.

pub(crate) fn check_federation_network( info: &FederationInfo, network: Network, ) -> Result<()>

Verifies that the supplied network matches the Bitcoin network in the connected client’s configuration.

pub async fn get_lightning_context( &self, ) -> Result<LightningContext, LightningRpcError>

Checks the Gateway’s current state and returns the proper LightningContext if it is available. Sometimes the lightning node will not be connected and this will return an error.

pub async fn unannounce_from_all_federations(&self)

Iterates through all of the federations the gateway is registered with and requests to remove the registration record.

impl Gateway

pub(crate) async fn public_key_v2( &self, federation_id: &FederationId, ) -> Option<PublicKey>

Retrieves the PublicKey of the Gateway module for a given federation for LNv2. This is NOT the same as the gateway_id, it is different per-connected federation.

pub async fn routing_info_v2( &self, federation_id: &FederationId, ) -> Option<RoutingInfo>

Returns payment information that LNv2 clients can use to instruct this Gateway to pay an invoice or receive a payment.

pub async fn select_client_v2( &self, federation_id: FederationId, ) -> Result<ClientHandleArc>

pub(crate) async fn send_payment_v2( &self, payload: SendPaymentPayload, ) -> Result<Result<[u8; 32], Signature>>

Instructs this gateway to pay a Lightning network invoice via the LNv2 protocol.

pub(crate) async fn create_bolt11_invoice_v2( &self, payload: CreateBolt11InvoicePayload, ) -> Result<Bolt11Invoice>

For the LNv2 protocol, this will create an invoice by fetching it from the connected Lightning node, then save the payment hash so that incoming HTLCs can be matched as a receive attempt to a specific federation.

pub async fn create_invoice_via_lnrpc_v2( &self, payment_hash: Hash, amount: Amount, description: Bolt11InvoiceDescription, expiry_time: u32, ) -> Result<Bolt11Invoice, String>

Retrieves a BOLT11 invoice from the connected Lightning node with a specific payment_hash.

pub async fn get_registered_incoming_contract_and_client_v2( &self, payment_hash: [u8; 32], amount_msats: u64, ) -> Result<(IncomingContract, ClientHandleArc)>

Retrieves the persisted CreateInvoicePayload from the database specified by the payment_hash and the ClientHandleArc specified by the payload’s federation_id.