Crate ln_gateway

Modules

client

config 🔒

db 🔒

envs

federation_manager 🔒

gateway_lnrpc

gateway_module_v2

lightning

rpc

state_machine

types 🔒

Map gateway_lnrpc protobuf types to rust types

Structs

Gateway

GatewayParameters

GatewayParameters is a helper struct that can be derived from GatewayOpts that holds the CLI or environment variables that are specified by the user.

Enums

GatewayError

Errors that can occur while processing incoming HTLC’s, making outgoing payments, registering with connected federations, or responding to webserver requests.

GatewayState

ReceivePaymentStreamAction 🔒

The action to take after handling a payment stream.

Constants

DB_FILE 🔒

Name of the gateway’s database that is used for metadata and configuration storage.

DEFAULT_FEES

The default routing fees that the gateway charges for incoming and outgoing payments. Identical to the Lightning Network.

DEFAULT_MODULE_KINDS 🔒

The non-lightning default module types that the Gateway supports.

DEFAULT_NETWORK

Default Bitcoin network for testing purposes.

DEFAULT_NUM_ROUTE_HINTS 🔒

The default number of route hints that the legacy gateway provides for invoice creation.

EXPIRATION_DELTA_MINIMUM_V2 🔒

LNv2 CLTV Delta in blocks

GW_ANNOUNCEMENT_TTL 🔒

How long a gateway announcement stays valid

LDK_NODE_DB_FOLDER 🔒

Name of the folder that the gateway uses to store its node database when running in LDK mode.

Type Aliases

Result