Crate ln_gateway

Modules

client

db 🔒

envs

gateway_lnrpc

lightning

rpc

state_machine

types 🔒

Map gateway_lnrpc protobuf types to rust types

Structs

ClientsJoinLock 🔒

Gateway

GatewayOpts 🔒

GatewayParameters

GatewayParameters is a helper struct that can be derived from GatewayOpts that holds the CLI or environment variables that are specified by the user.

LightningContext

Represents an active connection to the lightning node.

PrettyInterceptHtlcRequest 🔒

Enums

GatewayError

GatewayState