Crate ln_gateway

Modules

client

db

gateway_lnrpc

lnd

lnrpc_client

rpc

state_machine

types

Map gateway_lnrpc protobuf types to rust types

utils

Structs

Gateway

GatewayOpts

GatewayParameters 🔒

GatewayParameters is a helper struct that can be derived from GatewayOpts that holds the CLI or environment variables that are specified by the user.

Enums

GatewayError

GatewayState