Trait ln_gateway::client::IGatewayClientBuilder

pub trait IGatewayClientBuilder: Debug {
    // Required methods
    fn build<'life0, 'async_trait>(
        &'life0 self,
        config: GatewayClientConfig,
        decoders: ModuleDecoderRegistry,
        module_gens: ModuleGenRegistry
    ) -> Pin<Box<dyn Future<Output = Result<Client<GatewayClientConfig>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_config<'life0, 'async_trait>(
        &'life0 self,
        connect: WsFederationConnect,
        mint_channel_id: u64,
        node_pubkey: PublicKey,
        module_gens: ModuleGenRegistry
    ) -> Pin<Box<dyn Future<Output = Result<GatewayClientConfig>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn save_config(&self, config: GatewayClientConfig) -> Result<()>;
    fn load_configs(&self) -> Result<Vec<GatewayClientConfig>>;
}

Trait for gateway federation client builders

Required Methods

fn build<'life0, 'async_trait>( &'life0 self, config: GatewayClientConfig, decoders: ModuleDecoderRegistry, module_gens: ModuleGenRegistry ) -> Pin<Box<dyn Future<Output = Result<Client<GatewayClientConfig>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Build a new gateway federation client

fn create_config<'life0, 'async_trait>( &'life0 self, connect: WsFederationConnect, mint_channel_id: u64, node_pubkey: PublicKey, module_gens: ModuleGenRegistry ) -> Pin<Box<dyn Future<Output = Result<GatewayClientConfig>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a new gateway federation client config from connect info

fn save_config(&self, config: GatewayClientConfig) -> Result<()>

Save and persist the configuration of the gateway federation client

fn load_configs(&self) -> Result<Vec<GatewayClientConfig>>

Load all gateway client configs from the work directory