ln_gateway
Module lightning

Modules

cln

extension

ldk

lnd

Structs

ChannelInfo

CloseChannelsWithPeerResponse

CreateInvoiceRequest

CreateInvoiceResponse

GatewayLightningBuilder

GetBalancesResponse

GetLnOnchainAddressResponse

GetNodeInfoResponse

GetRouteHintsRequest

GetRouteHintsResponse

InterceptPaymentRequest

InterceptPaymentResponse

LightningContext

Represents an active connection to the lightning node.

ListActiveChannelsResponse

OpenChannelResponse

PayInvoiceRequest

PayInvoiceResponse

PayPrunedInvoiceRequest

WithdrawOnchainResponse

Enums

InvoiceDescription

LightningMode

LightningRpcError

PaymentAction

Constants

MAX_LIGHTNING_RETRIES

Traits

ILnRpcClient

A trait that the gateway uses to interact with a lightning node. This allows the gateway to be agnostic to the specific lightning node implementation being used.

LightningBuilder

Type Aliases

RouteHtlcStream