Trait ln_gateway::lnrpc_client::ILnRpcClient

pub trait ILnRpcClient: Debug + Send + Sync {
    // Required methods
    fn info<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<GetNodeInfoResponse, LightningRpcError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn routehints<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<GetRouteHintsResponse, LightningRpcError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn pay<'life0, 'async_trait>(
        &'life0 self,
        invoice: PayInvoiceRequest
    ) -> Pin<Box<dyn Future<Output = Result<PayInvoiceResponse, LightningRpcError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn route_htlcs<'a, 'life0, 'async_trait>(
        self: Box<Self>,
        task_group: &'life0 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<(RouteHtlcStream<'a>, Arc<dyn ILnRpcClient>), LightningRpcError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait;
    fn complete_htlc<'life0, 'async_trait>(
        &'life0 self,
        htlc: InterceptHtlcResponse
    ) -> Pin<Box<dyn Future<Output = Result<EmptyResponse, LightningRpcError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn info<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<GetNodeInfoResponse, LightningRpcError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get the public key and alias of the lightning node

fn routehints<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Result<GetRouteHintsResponse, LightningRpcError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Get route hints to the lightning node

fn pay<'life0, 'async_trait>( &'life0 self, invoice: PayInvoiceRequest ) -> Pin<Box<dyn Future<Output = Result<PayInvoiceResponse, LightningRpcError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Attempt to pay an invoice using the lightning node

fn route_htlcs<'a, 'life0, 'async_trait>( self: Box<Self>, task_group: &'life0 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<(RouteHtlcStream<'a>, Arc<dyn ILnRpcClient>), LightningRpcError>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait,

fn complete_htlc<'life0, 'async_trait>( &'life0 self, htlc: InterceptHtlcResponse ) -> Pin<Box<dyn Future<Output = Result<EmptyResponse, LightningRpcError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,