Module ln_gateway::state_machine::pay

Structs

GatewayPayCancelContract

GatewayPayClaimOutgoingContract

GatewayPayCommon

GatewayPayInvoice

GatewayPayStateMachine

GatewayPayWaitForSwapPreimage

OutgoingPaymentError

PaymentParameters 🔒

Enums

GatewayPayStates

State machine that executes the Lightning payment on behalf of the fedimint user that requested an invoice to be paid.

OutgoingContractError

OutgoingPaymentErrorType