Struct ln_gateway::Gateway

pub struct Gateway {
    pub(crate) lightning_builder: Arc<dyn LightningBuilder + Send + Sync>,
    pub(crate) gateway_parameters: GatewayParameters,
    pub state: Arc<RwLock<GatewayState>>,
    pub(crate) client_builder: GatewayClientBuilder,
    pub(crate) gateway_db: Database,
    pub(crate) clients: FederationToClientMap,
    pub(crate) scid_to_federation: ScidToFederationMap,
    pub gateway_id: PublicKey,
    pub(crate) channel_id_generator: Arc<Mutex<u64>>,
}

Fields
lightning_builder: Arc<dyn LightningBuilder + Send + Sync>gateway_parameters: GatewayParametersstate: Arc<RwLock<GatewayState>>client_builder: GatewayClientBuildergateway_db: Databaseclients: FederationToClientMapscid_to_federation: ScidToFederationMapgateway_id: PublicKeychannel_id_generator: Arc<Mutex<u64>>
Implementations

impl Gateway

pub async fn new_with_custom_registry( lightning_builder: Arc<dyn LightningBuilder + Send + Sync>, client_builder: GatewayClientBuilder, listen: SocketAddr, api_addr: SafeUrl, cli_password: Option<String>, network: Option<Network>, fees: RoutingFees, num_route_hints: u32, gateway_db: Database ) -> Result<Gateway>

pub async fn new_with_default_modules() -> Result<Gateway>

pub async fn get_gateway_id(gateway_db: Database) -> PublicKey

pub async fn dump_database<'a>( dbtx: &mut DatabaseTransaction<'_>, prefix_names: Vec<String> ) -> Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send>)> + 'a>

pub async fn run(self, tg: &mut TaskGroup) -> Result<TaskShutdownToken>

pub(crate) async fn start_webserver(&mut self, task_group: &mut TaskGroup)

pub(crate) async fn start_gateway( self, task_group: &mut TaskGroup ) -> Result<()>

pub(crate) async fn handle_disconnect(&mut self, htlc_task_group: TaskGroup)

pub async fn handle_htlc_stream( &self, stream: RouteHtlcStream<'_>, handle: TaskHandle )

pub(crate) async fn set_gateway_state(&mut self, state: GatewayState)

pub async fn handle_get_info(&self) -> Result<GatewayInfo>

pub async fn handle_get_federation_config( &self, federation_id: Option<FederationId> ) -> Result<GatewayFedConfig>

pub async fn handle_balance_msg( &self, payload: BalancePayload ) -> Result<Amount>

pub async fn handle_address_msg( &self, payload: DepositAddressPayload ) -> Result<Address>

pub async fn handle_withdraw_msg( &self, payload: WithdrawPayload ) -> Result<Txid>

pub(crate) async fn handle_pay_invoice_msg( &self, payload: PayInvoicePayload ) -> Result<Preimage>

pub(crate) async fn handle_connect_federation( &mut self, payload: ConnectFedPayload ) -> Result<FederationInfo>

pub async fn handle_leave_federation( &mut self, payload: LeaveFedPayload ) -> Result<()>

pub async fn handle_backup_msg(&self, __arg1: BackupPayload) -> Result<()>

pub async fn handle_restore_msg(&self, __arg1: RestorePayload) -> Result<()>

pub async fn handle_set_configuration_msg( &self, __arg1: SetConfigurationPayload ) -> Result<()>

pub(crate) async fn update_federation_routing_fees( &self, routing_fees: Option<String>, gateway_config: &GatewayConfiguration ) -> Result<()>

Updates the routing fees for every federation configuration. Also triggers a re-register of the gateway with all federations.

TODO: Once per-federation fees are supported, this function should only update a single federation’s routing fees at a time.

pub(crate) async fn register_all_federations( gateway: &Gateway, gateway_config: &GatewayConfiguration )

Iterates through all of the federation configurations and registers the gateway with each federation.

pub async fn get_gateway_configuration(&self) -> Option<GatewayConfiguration>

This function will return a GatewayConfiguration one of two ways. To avoid conflicting configs, the below order is the order in which the gateway will respect configurations:

GatewayConfiguration is read from the database.

All cli or environment variables are set such that we can create a GatewayConfiguration

pub async fn remove_client( &self, federation_id: FederationId ) -> Result<ClientArc>

pub async fn select_client( &self, federation_id: FederationId ) -> Result<ClientArc>

pub(crate) async fn load_clients(&mut self)

pub(crate) async fn register_clients_timer( &mut self, task_group: &mut TaskGroup )

pub(crate) async fn fetch_lightning_route_hints_try( lnrpc: &dyn ILnRpcClient, num_route_hints: u32 ) -> Result<Vec<RouteHint>>

pub(crate) async fn fetch_lightning_route_hints( lnrpc: Arc<dyn ILnRpcClient>, num_route_hints: u32 ) -> Result<Vec<RouteHint>>

pub(crate) async fn make_federation_info( &self, client: &ClientArc, federation_id: FederationId ) -> FederationInfo

pub(crate) async fn check_federation_network( &self, info: &FederationInfo, network: Network ) -> Result<()>

pub async fn get_lightning_context( &self ) -> Result<LightningContext, LightningRpcError>

Checks the Gateway’s current state and returns the proper LightningContext if it is available. Sometimes the lightning node will not be connected and this will return an error.